Handle missing data location in logfile

+++ biomero/slurm_client.py
@@ -667,13 +667,17 @@ class SlurmClient(Connection):
         # data
         if data_location is None:
             data_location = self.extract_data_location_from_log(logfile)
-        rmdata = f"rm -rf {data_location} {data_location}.*"
-        cmds.append(rmdata)
-        
-        # convert config file
-        config_file = f"config_{os.path.basename(data_location)}.txt"
-        rmconfig = f"rm {config_file}"
-        cmds.append(rmconfig)
+            
+        if data_location:
+            rmdata = f"rm -rf {data_location} {data_location}.*"
+            cmds.append(rmdata)
+            
+            # convert config file
+            config_file = f"config_{os.path.basename(data_location)}.txt"
+            rmconfig = f"rm {config_file}"
+            cmds.append(rmconfig)
+        else:
+            logger.warning(f"Could not extract data location from log {logfile}. Skipping cleanup.")
 
         try:
             # do as much as possible, not conditional removal
